Draw lines between stars in constellations
This is the minimum effort implementation for drawing a set of lines
between points that do not intersect: simply picking a single center at
random and radiating lines to all other points in the set.

Also renames the initial setup variable `constellations` to
`constellationZones` to avoid confusion with the new member variable.